requirements.txt
pandas
numpy
matplotlib
scikit-learn
tensorflow
streamlit
fastapi
uvicorn
joblib
python-jose
PyJWT
